+++ web/src/App.tsx
@@ -9,6 +9,7 @@ import Review from "./screens/Review";
 import Decks from "./screens/Decks";
 import Deck from "./screens/Deck";
 import MyDeck from "./screens/MyDeck";
+import Hardest from "./screens/Hardest";
 import Drill from "./screens/Drill";
 import GroupBoard from "./screens/GroupBoard";
 import Comprehension from "./screens/Comprehension";
@@ -56,6 +57,7 @@ export default function App() {
           <Route path="/grammar" element={<Grammar />} />
           <Route path="/vocab" element={<Decks />} />
           <Route path="/my-deck" element={<MyDeck />} />
+          <Route path="/hardest" element={<Hardest />} />
           <Route path="/vocab/:level/:tag" element={<Deck />} />
           <Route path="/comprehension" element={<Comprehension />} />
           <Route path="/comprehension/:id" element={<ComprehensionSet />} />